feat(example): add communication example
