+++ client/examples/cli/main.rs
@@ -5,11 +5,8 @@ mod arguments;
 
 use arguments::*;
 use clap::Clap;
-use futures::{executor::block_on, StreamExt};
-use iota_stronghold::{home_dir, naive_kdf, Location, RecordHint, ResultMessage, StatusMessage, Stronghold};
-
-#[cfg(feature = "communication")]
-use iota_stronghold::Multiaddr;
+use futures::executor::block_on;
+use iota_stronghold::{home_dir, naive_kdf, Location, RecordHint, StatusMessage, Stronghold};
 
 use riker::actors::*;
 use std::error::Error;
@@ -314,88 +311,6 @@ fn take_ownership_command(
     todo!()
 }
 
-// /// Relays a request to a remote stronghold instance.
-// #[cfg(feature = "communication")]
-// fn relay_command(stronghold: &mut iota_stronghold::Stronghold, client_path: Vec<u8>) -> Result<(), Box<dyn Error>> {
-
-//     todo!()
-// }
-
-/// Returns a list of all available peers
-#[cfg(feature = "communication")]
-fn list_peers_command(stronghold: &mut iota_stronghold::Stronghold) -> Result<(), Box<dyn Error>> {
-    match block_on(stronghold.get_swarm_info()) {
-        ResultMessage::Ok((_, _, peers)) => {
-            let info = format!(
-                r#"
-            Peers
-            ===
-            {:?}
-            "#,
-                peers
-            );
-            println!("{}", info)
-        }
-        ResultMessage::Error(e) => return Err(Box::from(format!("{}", e))),
-    }
-
-    Ok(())
-}
-
-/// Displays the swarm info of this stronghold instance
-#[cfg(feature = "communication")]
-fn show_swarm_info_command(stronghold: &mut iota_stronghold::Stronghold) -> Result<(), Box<dyn Error>> {
-    match block_on(stronghold.get_swarm_info()) {
-        ResultMessage::Ok((peer_id, addresses, peers)) => {
-            let info = format!(
-                r#"
-                Swarm Info:
-                ===
-                Peer Id : {},
-                Addresses: {:?},
-                Peers: {:?}
-            "#,
-                peer_id, addresses, peers
-            );
-
-            println!("{}", info)
-        }
-        ResultMessage::Error(e) => return Err(Box::from(format!("{}", e))),
-    }
-    Ok(())
-}
-
-fn start_listening_command(
-    multiaddr: &str,
-    stronghold: &mut iota_stronghold::Stronghold,
-    client_path: Vec<u8>,
-) -> Result<(), Box<dyn Error>> {
-    let multiaddress: Multiaddr = multiaddr.parse()?;
-
-    // spawn communication actor
-    let comm = stronghold.spawn_communication();
-
-    let result = block_on(stronghold.start_listening(Some(multiaddress)));
-
-    println!("Listening on addr: {:?}", result);
-
-    let (tx, rx): (
-        futures::channel::mpsc::Sender<usize>,
-        futures::channel::mpsc::Receiver<usize>,
-    ) = futures::channel::mpsc::channel(1);
-
-    let waiter = async {
-        let fut = rx.map(|f| f).collect::<Vec<usize>>();
-
-        fut.await
-    };
-
-    block_on(waiter);
-
-    // todo!()
-    Ok(())
-}
-
 fn main() -> Result<(), Box<dyn Error>> {
     let app = ExampleApp::parse();
     let system = ActorSystem::new().expect(line_error!());
@@ -445,19 +360,6 @@ fn main() -> Result<(), Box<dyn Error>> {
         Commands::Delete { record_path, pass } => {
             delete_from_store_command(pass.as_str(), record_path.as_str(), &mut stronghold, client_path)
         }
-
-        #[cfg(feature = "communication")]
-        Commands::Relay { id, path } => {
-            todo!()
-        }
-        #[cfg(feature = "communication")]
-        Commands::Peers {} => list_peers_command(&mut stronghold),
-
-        #[cfg(feature = "communication")]
-        Commands::SwarmInfo {} => show_swarm_info_command(&mut stronghold),
-
-        #[cfg(feature = "communication")]
-        Commands::Listen { multiaddr } => start_listening_command(multiaddr.as_str(), &mut stronghold, client_path),
     }?;
 
     Ok(())
