+++ client/examples/comm/README.md
@@ -1,0 +1,7 @@
+# Stronghold Communication Example
+
+[TODO]
+
+## Quickstart
+
+[TODO]
